no single page pdf download in double page view
Some users were confused when clicking on the single page pdf download
icon when the double page view was active. The PDF didn't reflect the
view the user had when clicking. To prevent this confusion in the future
the button is hidden now as image displaying and PDF generation are two
independend features.

fixes #15622