Updates MariaDB Java Client and persistece.xml changes
We had the problem, that the database scheme was not updated
automatically when using a version of the mariadb-java-client newer than
2.3.0.
This seems to be a problem related to EclipseLink not supporting MariaDB
officially until now. The proposed workaround is to set the
target-database explicitly to "MySQL". See here for further details:
* https://jira.mariadb.org/browse/CONJ-702